@@ -1,6 +1,6 @@
 __all__ = [ "__version__" ]  # pylint: disable=unused-variable
 
 __version__ = {}
-__version__["short"] = "1.0.3"
+__version__["short"] = "1.0.4"
 __version__["tag"] = "stable"
 __version__["full"] = f"{__version__['short']}-{__version__['tag']}"
